+++ src/ExceptionFilter/ExceptionFilterInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sirix\SentryPsr\ExceptionFilter;
+
+use Throwable;
+
+interface ExceptionFilterInterface
+{
+    public function shouldCapture(Throwable $throwable, ExceptionFilterContext $exceptionFilterContext): bool;
+}