+++ src/Listener/SentryCommandListenerFactory.php
@@ -10,6 +10,7 @@ use Sentry\State\HubInterface;
 use Sirix\ContainerResolver\ConfigReader;
 use Sirix\ContainerResolver\ContainerResolver;
 use Sirix\SentryPsr\Config\SentryPsrConfig;
+use Sirix\SentryPsr\ExceptionFilter\ExceptionFilterResolver;
 use Sirix\SentryPsr\Helper\LoggerHelper;
 use Sirix\SentryPsr\Lifecycle\SentryLifecycle;
 use Sirix\SentryPsr\Redaction\SentryRedactorFactory;
@@ -34,6 +35,7 @@ class SentryCommandListenerFactory
             logger: LoggerHelper::getLogger($containerResolver),
             sentryLifecycle: $containerResolver->get(SentryLifecycle::class),
             redactor: (new SentryRedactorFactory())($container),
+            exceptionFilter: ExceptionFilterResolver::fromConfiguredContainer($containerResolver, $configReader),
         );
     }
 }